refactor(token): move HTTP OAuth middleware into token pkg
Asmir mentioned in review that we use this in a number of places,
move it down & add a "add to request" method for populating the client
side of request construction
