+++ auth/token/context.go
@@ -2,6 +2,11 @@ package token
 
 import (
 	"context"
+	"fmt"
+	"net/http"
+	"strings"
+
+	"github.com/qri-io/qri/api/util"
 )
 
 // CtxKey defines a distinct type for context keys used by the access
@@ -24,4 +29,49 @@ func FromCtx(ctx context.Context) string {
 		return s
 	}
 	return ""
+}
+
+const (
+	// httpAuthorizationHeader is the http header field to check for tokens,
+	// follows OAuth 2.0 spec
+	httpAuthorizationHeader = "authorization"
+	// httpAuthorizationBearerPrefix is a prefix before a token in the
+	// Authorization header field. Follows OAuth 2.0 spec
+	httpAuthorizationBearerPrefix = "Bearer "
+)
+
+// OAuthTokenMiddleware parses any "authorization" header containing a Bearer
+// token & adds it to the request context
+func OAuthTokenMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		reqToken := r.Header.Get(httpAuthorizationHeader)
+		if reqToken == "" && r.FormValue(httpAuthorizationHeader) != "" {
+			reqToken = r.FormValue(httpAuthorizationHeader)
+		}
+		if reqToken == "" {
+			next.ServeHTTP(w, r)
+			return
+		}
+
+		if !strings.HasPrefix(reqToken, httpAuthorizationBearerPrefix) {
+			util.WriteErrResponse(w, http.StatusBadRequest, fmt.Errorf("bad token"))
+			return
+		}
+
+		tokenStr := strings.TrimPrefix(reqToken, httpAuthorizationBearerPrefix)
+		ctx := AddToContext(r.Context(), tokenStr)
+
+		r = r.WithContext(ctx)
+		next.ServeHTTP(w, r)
+	})
+}
+
+// AddContextTokenToRequest checks the supplied context for an auth token and
+// adds it to an http request, returns true if a token is added
+func AddContextTokenToRequest(ctx context.Context, r *http.Request) (*http.Request, bool) {
+	if s := FromCtx(ctx); s != "" {
+		r.Header.Set(httpAuthorizationHeader, strings.Join([]string{httpAuthorizationBearerPrefix, s}, ""))
+		return r, true
+	}
+	return r, false
 }
